Actors with type classes. Tests added.

+++ src/main/scala/com/datio/akkatraining/json/BeerRequest.scala
@@ -1,7 +1,7 @@
 package com.datio.akkatraining.json
 
 case class BeerRequest(client: String,
-                       number: Long,
+                       number: Int,
                        trade: String)
 
 case class ClientInfo(defaulter: Boolean, reason: String)
